+++ include/pybind11/iostream.h
@@ -6,17 +6,19 @@
 #include <ostream>
 #include <string>
 #include <memory>
+#include <iostream>
 
 NAMESPACE_BEGIN(PYBIND11_NAMESPACE)
-
 NAMESPACE_BEGIN(detail)
 
+// Buffer that writes to Python instead of C++
 class pythonbuf : public std::streambuf {
 private:
-    typedef std::streambuf::traits_type traits_type;
+    using traits_type = std::streambuf::traits_type;
     char          d_buffer[1024];
 
-    object pyostream;
+    object pywrite;
+    object pyflush;
 
     int overflow(int c) {
         if (!traits_type::eq_int_type(c, traits_type::eof())) {
@@ -28,11 +30,11 @@ private:
 
     int sync() {
         if (this->pbase() != this->pptr()) {
-            std::string line(this->pbase(), this->pptr());
+            // This subtraction cannot be negative, so dropping the sign
+            str line(this->pbase(), static_cast<size_t>(this->pptr() - this->pbase()));
 
-            auto write = pyostream.attr("write");
-            write(line);
-            pyostream.attr("flush")();
+            pywrite(line);
+            pyflush();
 
             this->setp(this->pbase(), this->epptr());
         }
@@ -40,11 +42,13 @@ private:
     }
 public:
     pythonbuf(object pyostream)
-        : pyostream(pyostream) {
+        : pywrite(pyostream.attr("write")),
+          pyflush(pyostream.attr("flush")) {
         this->setp(this->d_buffer, this->d_buffer + sizeof(this->d_buffer) - 1);
     }
 };
 
+// Stream that writes to Python instead of C++ file descriptors
 class opythonstream : private virtual pythonbuf, public std::ostream {
 public:
     opythonstream(object pyostream)
@@ -56,41 +60,132 @@ public:
 
 NAMESPACE_END(detail)
 
+
 /** \rst
-    Scope redirect
-    This a move-only guard and only a single instance can exist.
+    Scoped ostream output redirect
+    This a move-only guard that redirects output.
 
     .. code-block:: cpp
 
         #include <pybind11/iostream.h>
 
         int main() {
-            py::scoped_output_redirect redirect{};
+            py::scoped_ostream_redirect output;
             std::cout << "Hello, World!"; // Python stdout
         } // <-- return std::cout to normal
 
-    You can use py::ios::stderr as the constructor argument to gaurd stderr instead.
+    You can explicitly pass the c++ stream and the python object, for example to gaurd stderr instead.
+
+    .. code-block:: cpp
+        int main() {
+            py::scoped_ostream_redirect output{std::cerr, py::module::import("sys").attr("stderr")};
+            std::cerr << "Hello, World!";
+        }
  \endrst */
 
-class scoped_output_redirect {
+class scoped_ostream_redirect {
     std::streambuf * old {nullptr};
     std::ostream& costream;
     detail::opythonstream buffer;
 
 public:
-    scoped_output_redirect( std::ostream& costream, object pyostream) : costream(costream), buffer(pyostream) {
+    scoped_ostream_redirect(
+            std::ostream& costream = std::cout,
+            object pyostream = module::import("sys").attr("stdout") )
+        : costream(costream), buffer(pyostream) {
         old = costream.rdbuf(buffer.rdbuf());
     }
 
-    ~scoped_output_redirect() {
+    ~scoped_ostream_redirect() {
         costream.rdbuf(old);
     }
 
-    scoped_output_redirect(const scoped_output_redirect &) = delete;
-    scoped_output_redirect(scoped_output_redirect &&other) = default;
-    scoped_output_redirect &operator=(const scoped_output_redirect &) = delete;
-    scoped_output_redirect &operator=(scoped_output_redirect &&) = delete;
+    scoped_ostream_redirect(const scoped_ostream_redirect &) = delete;
+    scoped_ostream_redirect(scoped_ostream_redirect &&other) = default;
+    scoped_ostream_redirect &operator=(const scoped_ostream_redirect &) = delete;
+    scoped_ostream_redirect &operator=(scoped_ostream_redirect &&) = delete;
 };
 
+NAMESPACE_BEGIN(detail)
+
+// Class to redirect output as a context manager. C++ backend.
+class OstreamRedirect {
+    bool do_stdout_;
+    bool do_stderr_;
+    std::unique_ptr<scoped_ostream_redirect> redirect_stdout;
+    std::unique_ptr<scoped_ostream_redirect> redirect_stderr;
+
+public:
+
+    OstreamRedirect(bool do_stdout = false, bool do_stderr = false)
+        : do_stdout_(do_stdout), do_stderr_(do_stderr) {}
+
+    void enter() {
+        // If stdout is true, or if both are false
+        if (do_stdout_ || (!do_stdout_ && !do_stderr_)) {
+            redirect_stdout.reset(
+                new scoped_ostream_redirect(
+                    std::cout,
+                    module::import("sys").attr("stdout")
+                )
+            );
+        }
+
+        if (do_stderr_) {
+            redirect_stderr.reset(
+                new scoped_ostream_redirect(
+                    std::cerr,
+                    module::import("sys").attr("stderr")
+                )
+            );
+        }
+    }
+
+    void exit() {
+        redirect_stdout.reset();
+        redirect_stderr.reset();
+    }
+};
+
+NAMESPACE_END(detail)
+
+/** \rst
+    This is a helper function to add a C++ redirect context manager to Python instead of using a C++ guard.
+    To use it, add the following to your binding code:
+
+    .. code-block:: cpp
+
+        #include <pybind11/iostream.h>
+
+        ...
+
+        py::add_ostream_redirect(m, "ostream_redirect");
+
+    You now have a python context manager that redirects your output:
+
+    .. code-block:: python
+
+        with m.ostream_redirect():
+            m.print_to_cout_function()
+
+    This manager can optionally be told which streams to operate on:
+
+    .. code-block:: python
+
+        with m.ostream_redirect(stdout=true, stderr=true):
+            m.noisy_function_with_error_printing()
+
+ \endrst */
+
+class_<detail::OstreamRedirect> add_ostream_redirect(module m, std::string name = "ostream_redirect") {
+    return class_<detail::OstreamRedirect>(m, name.c_str())
+        .def(init<bool,bool>(), arg("stdout")=false, arg("stderr")=false)
+        .def("__enter__", [](detail::OstreamRedirect &self) {
+            self.enter();
+        })
+        .def("__exit__", [](detail::OstreamRedirect &self, args) {
+            self.exit();
+        });
+}
 
 NAMESPACE_END(PYBIND11_NAMESPACE)